+++ docs/release-notes.rst
@@ -1,9 +1,12 @@
 Release Notes
 =============
 
-Next release
-------------
+Unstable
+--------
 
+    * Add support for `PUBSUB` subcommands `CHANNELS`, `NUMSUB [arg] [args...]` and `NUMPAT`.
+    * Add method `set_result_callback(command, callback)` allowing the default reply callbacks to be changed, in the same way `set_response_callback(command, callback)` inherited from Redis-Py does for responses.
+    * Node manager now honors defined max_connections variable so connections that is emited from that class uses the same variable.
     * Add SSLClusterConnection for connecting over TLS/SSL to Redis Cluster
 
 1.3.3 (Dec 15, 2016)